add explanation to changes
Created Backup and Export classes to facilitate data backup and export functionality.

The Backup class provides methods to perform and restore backups from a source directory to a destination directory.

The Export class allows exporting data to a specified file in a specified format and importing data from a file.